pyproject.toml
[tool.poetry]
name = "mopblacklistbot"
version = "0.1.0"
description = "Discord Blacklist Bot for Tauri WoW"
authors = ["PederHA <peder.andresen@gmail.com>"]
readme= "README.md"

[tool.poetry.dependencies]
python = "^3.7"
"discord.py" = "^1.3.3"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
